30-12-2023

+++ app/src/main/java/com/example/appmonedero/data/model/Account.kt
@@ -1,3 +1,3 @@
 package com.example.appmonedero.data.model
 
-data class Account(var countryCode: CountryCode, var entidad:Int, var oficina: Int, var dc: Int, var accountNum: Int)
+data class Account(var countryCode: CountryCode, var entidad:Int, var oficina: Int, var dc: Int, var accountNum: Long)
